requirements.txt
appdirs==1.4.3
Django==1.11
gunicorn==19.7.1
kafka-python==1.3.3
packaging==16.8
pyparsing==2.2.0
pytz==2017.2
six==1.10.0
